move the image container to own file
