+++ fronts-client/src/components/inputs/InputImage.tsx
@@ -31,73 +31,11 @@ import {
   portraitCardImageCriteria,
 } from 'constants/image';
 import ImageDragIntentIndicator from 'components/image/ImageDragIntentIndicator';
+import { ImageInputImageContainer as ImageContainer } from 'components/image/ImageInputImageContainer';
 import { EditMode } from 'types/EditMode';
 import { selectEditMode } from '../../selectors/pathSelectors';
 import CircularIconContainer from '../icons/CircularIconContainer';
 import { error } from '../../styleConstants';
-
-const PORTRAIT_RATIO = 5 / 4;
-const NORMAL_PORTRAIT_WIDTH = 160;
-const SMALL_PORTRAIT_WIDTH = 60;
-const TEXTINPUT_HEIGHT = 30;
-
-const smallPortaitStyle = `
-  width: ${SMALL_PORTRAIT_WIDTH}px;
-  height: ${Math.floor(SMALL_PORTRAIT_WIDTH * PORTRAIT_RATIO)}px;
-  padding: 40% 0;
-  min-width: 50px;
-  margin: 0 auto;
-`;
-
-const normalPortraitStyle = `
-  width: ${NORMAL_PORTRAIT_WIDTH}px;
-  height: ${Math.floor(
-    NORMAL_PORTRAIT_WIDTH * PORTRAIT_RATIO + TEXTINPUT_HEIGHT
-  )}px;
-  `;
-
-const smallLandscapeStyle = `
-  width: 100%;
-  maxWidth: 180px;
-  height: 0;
-  padding: 40%;
-  minWidth: 50px;
-`;
-
-const normalLandscapeStyle = `
-  width: 100%;
-  maxWidth: 180px;
-  height: 115px;
-`;
-
-const getVariableImageContainerStyle = ({
-  portrait = false,
-  small = false,
-}: {
-  small?: boolean;
-  portrait?: boolean;
-}) =>
-  portrait
-    ? small
-      ? smallPortaitStyle
-      : normalPortraitStyle
-    : small
-    ? smallLandscapeStyle
-    : normalLandscapeStyle;
-
-// assuming any portrait image (ie height>width)
-// is in the 4:5 ratio for purposes of styling
-// the image container
-const ImageContainer = styled.div<{
-  small?: boolean;
-  portrait?: boolean;
-}>`
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  transition: background-color 0.15s;
-  ${getVariableImageContainerStyle}
-`;
 
 const AddImageButton = styled(ButtonDefault)<{ small?: boolean }>`
   background-color: ${({ small }) =>
